Conflicts:
	opencga-storage/opencga-storage-core/src/main/java/org/opencb/opencga/storage/core/alignment/tasks/AlignmentRegionCoverageCalculatorTask.java
	opencga-storage/opencga-storage-mongodb/src/main/java/org/opencb/opencga/storage/mongodb/alignment/IndexedAlignmentDBAdaptor.java
	opencga-storage/opencga-storage-mongodb/src/test/java/org/opencb/opencga/storage/mongodb/alignment/IndexedAlignmentDBAdaptorTest.java
	pom.xml

+++ opencga-storage/opencga-storage-mongodb/src/main/java/org/opencb/opencga/storage/mongodb/variant/VariantSourceMongoDBAdaptor.java
@@ -260,7 +260,7 @@ public class VariantSourceMongoDBAdaptor implements VariantSourceDBAdaptor {
                 .append(DBObjectToVariantSourceConverter.NUMPASSFILTERS_FIELD, global.getPassCount())
                 .append(DBObjectToVariantSourceConverter.NUMTRANSITIONS_FIELD, global.getTransitionsCount())
                 .append(DBObjectToVariantSourceConverter.NUMTRANSVERSIONS_FIELD, global.getTransversionsCount())
-                .append(DBObjectToVariantSourceConverter.MEANQUALITY_FIELD, (double) global.getMeanQuality());
+                .append(DBObjectToVariantSourceConverter.MEANQUALITY_FIELD, global.getMeanQuality());
 
         DBObject find = new BasicDBObject(DBObjectToVariantSourceConverter.FILEID_FIELD, variantSourceStats.getFileId())
                 .append(DBObjectToVariantSourceConverter.STUDYID_FIELD, variantSourceStats.getStudyId());
